request and response headers are back integers now

+++ init-mqtt.h
@@ -143,8 +143,8 @@ bool tryConnectMqtt(const char* server, int port, const char* user, const char* 
     && connectMqtt(user, pwd);
 }
 
-int saveMqttConfig(JsonObject& json) {
-  int res = 0;
+uint8_t saveMqttConfig(JsonObject& json) {
+  uint8_t res = 0;
   
   if(
       json.success()
